requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
twilio>=8.5.0
langchain>=0.1.0
openai>=1.0.0
pinecone-client>=2.0.0
redis>=5.0.0
aiohttp>=3.9.0
aiosmtplib>=2.0.0
asyncpg>=0.29.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
websockets>=11.0.0
python-multipart>=0.0.6
deepgram-sdk>=2.11.0
elevenlabs>=0.2.0
cachetools>=5.3.0
